+++ changelog.md
@@ -1,3 +1,8 @@
+## Release 0.1.3 2015/5/11
+* Removed gem resource
+* Fixed readme example
+* Minor fix to log resources in artifact resource
+* Fixed download\_path to display the location of the downloaded artifact
 ## Release 0.1.2 2016/5/10
 * Version bump for more minor format changes